initial Sqlite setup with context
