+++ Jetchat/app/src/main/java/com/example/compose/jetchat/Constants.kt
@@ -3,8 +3,8 @@ package com.example.compose.jetchat
 internal object Constants {
     internal const val OPENAI_TOKEN = "{OPENAI_KEY}"
 
-    /** Chat model: "gpt-4" or "gpt-3.5-turbo-0613" */
-    internal const val OPENAI_CHAT_MODEL = "gpt-3.5-turbo-0613"
+    /** Chat model: "gpt-4" or "gpt-3.5-turbo-16k" */
+    internal const val OPENAI_CHAT_MODEL = "gpt-3.5-turbo-16k"
 
     /** Embedding model: text-embedding-ada-002 */
     internal const val OPENAI_EMBED_MODEL = "text-embedding-ada-002"
